SES-16
Some fixes in Commutator and CommutatorManager

Small refactoring:
1. type for time interval has been changed from size_t to uiny32_t
2. MockedBaseModule doean't override onMessageReceived() and
attachToChammel() functions anymore
3. Utils::WaitingFor: sleep_for() call has been replaced with yield()
call

+++ Conveyor/Conveyor.cpp
@@ -16,7 +16,7 @@ void Conveyor::addLogicToChain(IAbstractLogicPtr pLogic)
   m_LogicChain.push_back(std::move(context));
 }
 
-void Conveyor::proceed(size_t nIntervalUs)
+void Conveyor::proceed(uint32_t nIntervalUs)
 {
   m_State.nCurrentTimeUs  += nIntervalUs;
   m_State.pSelectedLogic  = nullptr;
