+++ Network/UdpDispatcher.h
@@ -25,7 +25,7 @@ public:
   // overrides from IAbstractLogic interface
   uint16_t getStagesCount() override { return 1; }
   bool     prephareStage(uint16_t nStageId) override;
-  void     proceedStage(uint16_t nStageId, size_t nIntervalUs) override;
+  void     proceedStage(uint16_t nStageId, uint32_t nIntervalUs) override;
   size_t   getCooldownTimeUs() const override { return 3000; }
 
 private:
